Price prediction with multivariable regression
A series of exercises to learn how to make price predictions with a multivariable linear regression. I used the pre-loaded Boston house dataset included with Sklearn to work with. In this notebook, you can find a thorough analysis to understand which factors influence house pricing in the Boston housing market. Model evaluation and validation techniques were performed too.